Added Placeholders for UI components. Need to complete QueryParams
