+++ src/assets/bulma.scss
@@ -1,0 +1,24 @@
+
+@import '~bulma';
+
+
+
+// @import "bulma-0.9.1/sass/base/_all.sass";
+// @import "bulma-0.9.1/sass/helpers/_all.sass";
+// @import "bulma-0.9.1/sass/elements/_all.sass";
+// @import "bulma-0.9.1/sass/form/shared.sass";
+// @import "bulma-0.9.1/sass/form/input-textarea.sass";
+// @import "bulma-0.9.1/sass/form/tools.sass";
+// @import "bulma-0.9.1/sass/components/breadcrumb.sass";
+// @import "bulma-0.9.1/sass/components/dropdown.sass";
+// @import "bulma-0.9.1/sass/components/breadcrumb.sass";
+// @import "bulma-0.9.1/sass/components/level.sass";
+// @import "bulma-0.9.1/sass/components/menu.sass";
+// @import "bulma-0.9.1/sass/components/level.sass";
+// @import "bulma-0.9.1/sass/components/modal.sass";
+// @import "bulma-0.9.1/sass/components/navbar.sass";
+// @import "bulma-0.9.1/sass/components/pagination.sass";
+// @import "bulma-0.9.1/sass/components/panel.sass";
+// @import "bulma-0.9.1/sass/components/tabs.sass";
+// @import "bulma-0.9.1/sass/grid/_all.sass";
+// @import "bulma-0.9.1/sass/layout/_all.sass";
