+++ src/components/Header.vue
@@ -1,16 +1,6 @@
 <template>
     <nav class="navbar" role="navigation" aria-label="main navigation">
-    <div class="navbar-brand">
-        <a class="navbar-item" href="https://bulma.io">
-        <img src="https://bulma.io/images/bulma-logo.png" width="112" height="28">
-        </a>
 
-        <a role="button" class="navbar-burger" aria-label="menu" aria-expanded="false" data-target="navbarBasicExample">
-        <span aria-hidden="true"></span>
-        <span aria-hidden="true"></span>
-        <span aria-hidden="true"></span>
-        </a>
-    </div>
 
     <div id="navbarBasicExample" class="navbar-menu">
         <div class="navbar-start">
@@ -24,7 +14,8 @@
 
         <div class="navbar-item has-dropdown is-hoverable">
             <a class="navbar-link">
-            More
+            Links
+
             </a>
 
             <div class="navbar-dropdown">
@@ -72,6 +63,5 @@ export default {
 </script>
 
 <style lang="scss" scoped>
-    // @import 'bulma/css/bulma.css';
-    @import '~bulma';
+    @import '../assets/bulma.scss';
 </style>
